Analytics auto - refactored Analytics package
