+++ src/test/java/com/nci/testcases/Analytics_Test.java
@@ -13,13 +13,13 @@ import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
 import com.nci.Utilities.BrowserManager;
-import com.nci.clinicalTrial.pages.Analytics;
 import com.relevantcodes.extentreports.LogStatus;
+import gov.nci.WebAnalytics.AnalyticsLoad;
 
 public class Analytics_Test extends BaseClass {
 
 	// WebDriver driver;
-	Analytics analytics;
+	AnalyticsLoad analytics;
 
 	// ConfigReader config = new ConfigReader();
 
@@ -31,7 +31,7 @@ public class Analytics_Test extends BaseClass {
 		System.out.println("PageURL: " + pageURL);
 		driver = BrowserManager.startBrowser(browser, pageURL);
 		driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-		analytics = new Analytics(driver);
+		analytics = new AnalyticsLoad(driver);
 		System.out.println("Analytics setup done");
 	}
 
@@ -39,7 +39,7 @@ public class Analytics_Test extends BaseClass {
 	public void veriFySAccount() {
 		String sAccountBlob = analytics.getSAccountText();
 		//System.out.println("s_account element: " + sAccountBlob);
-		Assert.assertTrue(sAccountBlob.contains(Analytics.S_ACCOUNT));
+		Assert.assertTrue(sAccountBlob.contains(AnalyticsLoad.S_ACCOUNT));
 		logger.log(LogStatus.INFO, "s_account element: " + sAccountBlob);
 		logger.log(LogStatus.PASS, "The page contains the s_account variable.");		
 	}
